init: move slave core init to src/init.c
It's generic code and not xtensa specific.
