+++ src/arch/xtensa/init.c
@@ -108,57 +108,3 @@ int arch_init(void)
 	register_exceptions();
 	return 0;
 }
-
-#if CONFIG_MULTICORE
-
-#include <sof/debug/panic.h>
-#include <sof/drivers/idc.h>
-#include <sof/drivers/interrupt.h>
-#include <sof/lib/notifier.h>
-#include <sof/platform.h>
-#include <sof/schedule/edf_schedule.h>
-#include <sof/schedule/ll_schedule.h>
-#include <sof/schedule/ll_schedule_domain.h>
-#include <sof/schedule/schedule.h>
-#include <sof/schedule/task.h>
-#include <sof/spinlock.h>
-#include <sof/trace/trace.h>
-#include <ipc/trace.h>
-
-int slave_core_init(struct sof *sof)
-{
-	int err;
-
-	/* init architecture */
-	trace_point(TRACE_BOOT_ARCH);
-	err = arch_init();
-	if (err < 0)
-		panic(SOF_IPC_PANIC_ARCH);
-
-	trace_point(TRACE_BOOT_SYS_NOTIFIER);
-	init_system_notify(sof);
-
-	/* interrupts need to be initialized before any usage */
-	trace_point(TRACE_BOOT_PLATFORM_IRQ);
-	platform_interrupt_init();
-
-	trace_point(TRACE_BOOT_PLATFORM_SCHED);
-	scheduler_init_edf();
-	scheduler_init_ll(timer_domain_get());
-	scheduler_init_ll(dma_domain_get());
-
-	/* initialize IDC mechanism */
-	trace_point(TRACE_BOOT_PLATFORM_IDC);
-	err = idc_init();
-	if (err < 0)
-		return err;
-
-	trace_point(TRACE_BOOT_PLATFORM);
-
-	/* task initialized in edf_scheduler_init */
-	schedule_task(*task_main_get(), 0, UINT64_MAX);
-
-	return err;
-}
-
-#endif
